rename Description component to DescriptionPanel
