+++ src/components/DescriptionPanel.tsx
