+++ src/panel/description/module.ts
@@ -1,4 +1,4 @@
 import { PanelPlugin } from '@grafana/data';
-import { DescriptionPanel } from '../../components/Description';
+import { DescriptionPanel } from '../../components/DescriptionPanel';
 
 export const plugin = new PanelPlugin<{}>(DescriptionPanel);
